Address code review feedback: constants, safe split, correct tag cast
Agent-Logs-Url: https://github.com/MakSobo80/rpi/sessions/bf35f164-1cfb-41ee-b0e6-9e4276a51d63

+++ Notatnik/Notatnik/MainWindow.xaml.cs
@@ -143,7 +143,7 @@ namespace Notatnik
                     return;
                 }
 
-                int fileId = (int)(byte)selected.Tag;
+                int fileId = (int)selected.Tag;
                 var fileData = Database.GetFileById(fileId);
                 if (fileData == null)
                 {
